code cleanup and connection to db
