+++ templates/product_info.html
@@ -15,6 +15,7 @@
 <div class="container">
     <div class="row">
         <div class="col-md-7">
+            <p class="product-id" hidden>{{product._id}}</p>
             <h2 class="font2 mt-3 product-name" id="A{{product._id}}">{{product.productName}}</h2>
             <h5>Category: {{product.productCategory}}</h5>
             <p class="par1 mt-4">{{product.productDescription}}</p>
